refactor: changed the name of PillarsComponent to be PasCalCase for better consistency with other files
