+++ src/lib/components/PillarsComponent.svelte
